refactor(copilot): consolidate file tools into single e2b_file_tools.py
Delete file_tools.py and file_tools_test.py, extending e2b_file_tools.py
to handle both E2B (sandbox) and non-E2B (local SDK working dir) modes.

Each handler checks _get_sandbox(): if a sandbox exists, it routes to
the E2B filesystem; otherwise it falls back to the SDK working directory
with path validation, truncation detection, per-path edit locking, and
large content warnings.

All 104 tests from both files now live in e2b_file_tools_test.py.
tool_adapter.py imports exclusively from e2b_file_tools.
