+++ autogpt_platform/backend/backend/copilot/sdk/e2b_file_tools.py
@@ -1,8 +1,12 @@
-"""MCP file-tool handlers that route to the E2B cloud sandbox.
+"""Unified MCP file-tool handlers for both E2B (sandbox) and non-E2B (local) modes.
 
-When E2B is active, these tools replace the SDK built-in Read/Write/Edit/
-Glob/Grep so that all file operations share the same ``/home/user``
-and ``/tmp`` filesystems as ``bash_exec``.
+When E2B is active, Read/Write/Edit/Glob/Grep route to the sandbox so that
+all file operations share the same ``/home/user`` and ``/tmp`` filesystems
+as ``bash_exec``.
+
+In non-E2B mode (no sandbox), Read/Write/Edit operate on the SDK working
+directory (``/tmp/copilot-<session>/``), providing the same truncation
+detection and path-validation guarantees.
 
 SDK-internal paths (``~/.claude/projects/…/tool-results/``) are handled
 by the separate ``Read`` MCP tool registered in ``tool_adapter.py``.
@@ -36,6 +40,118 @@ logger = logging.getLogger(__name__)
 # decide whether the model is requesting the full file (and thus whether the
 # bridge copy is worthwhile).
 _DEFAULT_READ_LIMIT = 2000
+
+# Per-path lock for edit operations to prevent parallel lost updates.
+# When MCP tools are dispatched in parallel (readOnlyHint=True annotation),
+# two Edit calls on the same file could race through read-modify-write
+# and silently drop one change.  Keyed by resolved absolute path.
+_edit_locks: dict[str, asyncio.Lock] = {}
+
+# Inline content above this threshold triggers a warning — it survived this
+# time but is dangerously close to the API output-token truncation limit.
+_LARGE_CONTENT_WARN_CHARS = 50_000
+
+_READ_BINARY_EXTENSIONS = frozenset(
+    {
+        ".png",
+        ".jpg",
+        ".jpeg",
+        ".gif",
+        ".bmp",
+        ".ico",
+        ".webp",
+        ".pdf",
+        ".zip",
+        ".gz",
+        ".tar",
+        ".bz2",
+        ".xz",
+        ".7z",
+        ".exe",
+        ".dll",
+        ".so",
+        ".dylib",
+        ".bin",
+        ".o",
+        ".a",
+        ".pyc",
+        ".pyo",
+        ".class",
+        ".wasm",
+        ".mp3",
+        ".mp4",
+        ".avi",
+        ".mov",
+        ".mkv",
+        ".wav",
+        ".flac",
+        ".sqlite",
+        ".db",
+    }
+)
+
+
+def _is_likely_binary(path: str) -> bool:
+    """Heuristic check for binary files by extension."""
+    _, ext = os.path.splitext(path)
+    return ext.lower() in _READ_BINARY_EXTENSIONS
+
+
+_PARTIAL_TRUNCATION_MSG = (
+    "Your Write call was truncated (file_path missing but content "
+    "was present). The content was too large for a single tool call. "
+    "Write in chunks: use bash_exec with "
+    "'cat > file << \"EOF\"\\n...\\nEOF' for the first section, "
+    "'cat >> file << \"EOF\"\\n...\\nEOF' to append subsequent "
+    "sections, then reference the file with "
+    "@@agptfile:/path/to/file if needed."
+)
+
+_COMPLETE_TRUNCATION_MSG = (
+    "Your Write call had empty arguments — this means your previous "
+    "response was too long and the tool call was truncated by the API. "
+    "Break your work into smaller steps. For large content, write "
+    "section-by-section using bash_exec with "
+    "'cat > file << \"EOF\"\\n...\\nEOF' and "
+    "'cat >> file << \"EOF\"\\n...\\nEOF'."
+)
+
+_EDIT_PARTIAL_TRUNCATION_MSG = (
+    "Your Edit call was truncated (file_path missing but old_string/new_string "
+    "were present). The arguments were too large for a single tool call. "
+    "Break your edit into smaller replacements, or use bash_exec with "
+    "'sed' for large-scale find-and-replace."
+)
+
+
+def _check_truncation(file_path: str, content: str) -> dict[str, Any] | None:
+    """Return an error response if the args look truncated, else ``None``."""
+    if not file_path:
+        if content:
+            return _mcp(_PARTIAL_TRUNCATION_MSG, error=True)
+        return _mcp(_COMPLETE_TRUNCATION_MSG, error=True)
+    return None
+
+
+def _resolve_and_validate(
+    file_path: str, sdk_cwd: str
+) -> tuple[str, None] | tuple[None, dict[str, Any]]:
+    """Resolve *file_path* against *sdk_cwd* and validate it stays within bounds.
+
+    Returns ``(resolved_path, None)`` on success, or ``(None, error_response)``
+    on failure.
+    """
+    if not os.path.isabs(file_path):
+        resolved = os.path.normpath(os.path.join(sdk_cwd, file_path))
+    else:
+        resolved = os.path.normpath(file_path)
+
+    if not is_allowed_local_path(resolved, sdk_cwd):
+        return None, _mcp(
+            f"Path must be within the working directory: {os.path.basename(file_path)}",
+            error=True,
+        )
+    return resolved, None
 
 
 async def _check_sandbox_symlink_escape(
@@ -137,7 +253,7 @@ async def _sandbox_write(sandbox: Any, path: str, content: str | bytes) -> None:
 
 
 async def _handle_read_file(args: dict[str, Any]) -> dict[str, Any]:
-    """Read lines from a sandbox file, falling back to the local host for SDK-internal paths."""
+    """Read lines from a file — E2B sandbox, local SDK working dir, or SDK-internal paths."""
     if not args:
         return _mcp(
             "Your read_file call had empty arguments \u2014 this means your previous "
@@ -176,19 +292,54 @@ async def _handle_read_file(args: dict[str, Any]) -> dict[str, Any]:
                     result["content"][0]["text"] += annotation
         return result
 
-    result = _get_sandbox_and_path(file_path)
-    if isinstance(result, dict):
-        return result
-    sandbox, remote = result
+    sandbox = _get_sandbox()
+    if sandbox is not None:
+        # E2B path — read from sandbox filesystem
+        result = _get_sandbox_and_path(file_path)
+        if isinstance(result, dict):
+            return result
+        sandbox, remote = result
+
+        try:
+            raw: bytes = await sandbox.files.read(remote, format="bytes")
+            content = raw.decode("utf-8", errors="replace")
+        except Exception as exc:
+            return _mcp(f"Failed to read {remote}: {exc}", error=True)
+
+        lines = content.splitlines(keepends=True)
+        selected = list(itertools.islice(lines, offset, offset + limit))
+        numbered = "".join(
+            f"{i + offset + 1:>6}\t{line}" for i, line in enumerate(selected)
+        )
+        return _mcp(numbered)
+
+    # Non-E2B path — read from SDK working directory
+    sdk_cwd = get_sdk_cwd()
+    if not sdk_cwd:
+        return _mcp("No SDK working directory available", error=True)
+
+    resolved, err = _resolve_and_validate(file_path, sdk_cwd)
+    if err is not None:
+        return err
+    assert resolved is not None
+
+    if _is_likely_binary(resolved):
+        return _mcp(
+            f"Cannot read binary file: {os.path.basename(resolved)}. "
+            "Use bash_exec with 'xxd' or 'file' to inspect binary files.",
+            error=True,
+        )
 
     try:
-        raw: bytes = await sandbox.files.read(remote, format="bytes")
-        content = raw.decode("utf-8", errors="replace")
+        with open(resolved, encoding="utf-8", errors="replace") as f:
+            selected = list(itertools.islice(f, offset, offset + limit))
+    except FileNotFoundError:
+        return _mcp(f"File not found: {file_path}", error=True)
+    except PermissionError:
+        return _mcp(f"Permission denied: {file_path}", error=True)
     except Exception as exc:
-        return _mcp(f"Failed to read {remote}: {exc}", error=True)
+        return _mcp(f"Failed to read {file_path}: {exc}", error=True)
 
-    lines = content.splitlines(keepends=True)
-    selected = list(itertools.islice(lines, offset, offset + limit))
     numbered = "".join(
         f"{i + offset + 1:>6}\t{line}" for i, line in enumerate(selected)
     )
@@ -196,59 +347,100 @@ async def _handle_read_file(args: dict[str, Any]) -> dict[str, Any]:
 
 
 async def _handle_write_file(args: dict[str, Any]) -> dict[str, Any]:
-    """Write content to a sandbox file, creating parent directories as needed."""
+    """Write content to a file — E2B sandbox or local SDK working directory."""
     if not args:
-        return _mcp(
-            "Your write_file call had empty arguments \u2014 this means your previous "
-            "response was too long and the tool call was truncated by the API. "
-            "Break your work into smaller steps: use bash_exec with "
-            "'cat > file << \"EOF\"\\n...\\nEOF' for the first section, "
-            "'cat >> file << \"EOF\"\\n...\\nEOF' to append.",
-            error=True,
-        )
+        return _mcp(_COMPLETE_TRUNCATION_MSG, error=True)
     file_path: str = args.get("file_path", "")
     content: str = args.get("content", "")
 
-    if not file_path:
-        if content:
-            return _mcp(
-                "Partial truncation \u2014 file_path was cut off. "
-                "Write in chunks using bash_exec instead.",
-                error=True,
-            )
-        return _mcp("file_path is required", error=True)
+    truncation_err = _check_truncation(file_path, content)
+    if truncation_err is not None:
+        return truncation_err
 
-    result = _get_sandbox_and_path(file_path)
-    if isinstance(result, dict):
-        return result
-    sandbox, remote = result
+    sandbox = _get_sandbox()
+    if sandbox is not None:
+        # E2B path — write to sandbox filesystem
+        try:
+            remote = resolve_sandbox_path(file_path)
+        except ValueError as exc:
+            return _mcp(str(exc), error=True)
+
+        try:
+            parent = os.path.dirname(remote)
+            if parent and parent not in E2B_ALLOWED_DIRS:
+                await sandbox.files.make_dir(parent)
+            canonical_parent = await _check_sandbox_symlink_escape(sandbox, parent)
+            if canonical_parent is None:
+                return _mcp(
+                    f"Path must be within {E2B_ALLOWED_DIRS_STR}: {os.path.basename(parent)}",
+                    error=True,
+                )
+            remote = os.path.join(canonical_parent, os.path.basename(remote))
+            await _sandbox_write(sandbox, remote, content)
+        except Exception as exc:
+            return _mcp(f"Failed to write {remote}: {exc}", error=True)
+
+        msg = f"Successfully wrote to {remote}"
+        if len(content) > _LARGE_CONTENT_WARN_CHARS:
+            logger.warning(
+                "[Write] large inline content (%d chars) for %s",
+                len(content),
+                remote,
+            )
+            msg += (
+                f"\n\nWARNING: The content was very large ({len(content)} chars). "
+                "Next time, write large files in sections using bash_exec with "
+                "'cat > file << EOF ... EOF' and 'cat >> file << EOF ... EOF' "
+                "to avoid output-token truncation."
+            )
+        return _mcp(msg)
+
+    # Non-E2B path — write to SDK working directory
+    sdk_cwd = get_sdk_cwd()
+    if not sdk_cwd:
+        return _mcp("No SDK working directory available", error=True)
+
+    resolved, err = _resolve_and_validate(file_path, sdk_cwd)
+    if err is not None:
+        return err
+    assert resolved is not None
 
     try:
-        parent = os.path.dirname(remote)
-        if parent and parent not in E2B_ALLOWED_DIRS:
-            await sandbox.files.make_dir(parent)
-        canonical_parent = await _check_sandbox_symlink_escape(sandbox, parent)
-        if canonical_parent is None:
-            return _mcp(
-                f"Path must be within {E2B_ALLOWED_DIRS_STR}: {os.path.basename(parent)}",
-                error=True,
-            )
-        remote = os.path.join(canonical_parent, os.path.basename(remote))
-        await _sandbox_write(sandbox, remote, content)
+        parent = os.path.dirname(resolved)
+        if parent:
+            os.makedirs(parent, exist_ok=True)
+        with open(resolved, "w", encoding="utf-8") as f:
+            f.write(content)
     except Exception as exc:
-        return _mcp(f"Failed to write {remote}: {exc}", error=True)
+        logger.error("Write failed for %s: %s", resolved, exc, exc_info=True)
+        return _mcp(
+            f"Failed to write {os.path.basename(resolved)}: {type(exc).__name__}",
+            error=True,
+        )
 
-    return _mcp(f"Successfully wrote to {remote}")
+    msg = f"Successfully wrote to {resolved}"
+    if len(content) > _LARGE_CONTENT_WARN_CHARS:
+        logger.warning(
+            "[Write] large inline content (%d chars) for %s",
+            len(content),
+            resolved,
+        )
+        msg += (
+            f"\n\nWARNING: The content was very large ({len(content)} chars). "
+            "Next time, write large files in sections using bash_exec with "
+            "'cat > file << EOF ... EOF' and 'cat >> file << EOF ... EOF' "
+            "to avoid output-token truncation."
+        )
+    return _mcp(msg)
 
 
 async def _handle_edit_file(args: dict[str, Any]) -> dict[str, Any]:
-    """Replace a substring in a sandbox file, with optional replace-all support."""
+    """Replace a substring in a file — E2B sandbox or local SDK working directory."""
     if not args:
         return _mcp(
-            "Your edit_file call had empty arguments \u2014 this means your previous "
+            "Your Edit call had empty arguments \u2014 this means your previous "
             "response was too long and the tool call was truncated by the API. "
-            "Break your edit into smaller replacements, or use bash_exec with "
-            "'sed' for large-scale find-and-replace.",
+            "Break your work into smaller steps.",
             error=True,
         )
     file_path: str = args.get("file_path", "")
@@ -256,59 +448,118 @@ async def _handle_edit_file(args: dict[str, Any]) -> dict[str, Any]:
     new_string: str = args.get("new_string", "")
     replace_all: bool = args.get("replace_all", False)
 
+    # Partial truncation: file_path missing but edit strings present
     if not file_path:
         if old_string or new_string:
-            return _mcp(
-                "Your edit_file call was truncated (file_path missing but "
-                "old_string/new_string were present). Break your edit into "
-                "smaller replacements.",
-                error=True,
-            )
-        return _mcp("file_path is required", error=True)
+            return _mcp(_EDIT_PARTIAL_TRUNCATION_MSG, error=True)
+        return _mcp(
+            "Your Edit call had empty arguments \u2014 this means your previous "
+            "response was too long and the tool call was truncated by the API. "
+            "Break your work into smaller steps.",
+            error=True,
+        )
+
     if not old_string:
         return _mcp("old_string is required", error=True)
 
-    result = _get_sandbox_and_path(file_path)
-    if isinstance(result, dict):
-        return result
-    sandbox, remote = result
+    sandbox = _get_sandbox()
+    if sandbox is not None:
+        # E2B path — edit in sandbox filesystem
+        try:
+            remote = resolve_sandbox_path(file_path)
+        except ValueError as exc:
+            return _mcp(str(exc), error=True)
 
-    parent = os.path.dirname(remote)
-    canonical_parent = await _check_sandbox_symlink_escape(sandbox, parent)
-    if canonical_parent is None:
-        return _mcp(
-            f"Path must be within {E2B_ALLOWED_DIRS_STR}: {os.path.basename(parent)}",
-            error=True,
+        parent = os.path.dirname(remote)
+        canonical_parent = await _check_sandbox_symlink_escape(sandbox, parent)
+        if canonical_parent is None:
+            return _mcp(
+                f"Path must be within {E2B_ALLOWED_DIRS_STR}: {os.path.basename(parent)}",
+                error=True,
+            )
+        remote = os.path.join(canonical_parent, os.path.basename(remote))
+
+        try:
+            raw: bytes = await sandbox.files.read(remote, format="bytes")
+            content = raw.decode("utf-8", errors="replace")
+        except Exception as exc:
+            return _mcp(f"Failed to read {remote}: {exc}", error=True)
+
+        count = content.count(old_string)
+        if count == 0:
+            return _mcp(f"old_string not found in {file_path}", error=True)
+        if count > 1 and not replace_all:
+            return _mcp(
+                f"old_string appears {count} times in {file_path}. "
+                "Use replace_all=true or provide a more unique string.",
+                error=True,
+            )
+
+        updated = (
+            content.replace(old_string, new_string)
+            if replace_all
+            else content.replace(old_string, new_string, 1)
         )
-    remote = os.path.join(canonical_parent, os.path.basename(remote))
+        try:
+            await _sandbox_write(sandbox, remote, updated)
+        except Exception as exc:
+            return _mcp(f"Failed to write {remote}: {exc}", error=True)
 
-    try:
-        raw: bytes = await sandbox.files.read(remote, format="bytes")
-        content = raw.decode("utf-8", errors="replace")
-    except Exception as exc:
-        return _mcp(f"Failed to read {remote}: {exc}", error=True)
+        return _mcp(f"Edited {remote} ({count} replacement{'s' if count > 1 else ''})")
 
-    count = content.count(old_string)
-    if count == 0:
-        return _mcp(f"old_string not found in {file_path}", error=True)
-    if count > 1 and not replace_all:
-        return _mcp(
-            f"old_string appears {count} times in {file_path}. "
-            "Use replace_all=true or provide a more unique string.",
-            error=True,
+    # Non-E2B path — edit in SDK working directory
+    sdk_cwd = get_sdk_cwd()
+    if not sdk_cwd:
+        return _mcp("No SDK working directory available", error=True)
+
+    resolved, err = _resolve_and_validate(file_path, sdk_cwd)
+    if err is not None:
+        return err
+    assert resolved is not None
+
+    # Per-path lock prevents parallel edits from racing through
+    # the read-modify-write cycle and silently dropping changes.
+    if resolved not in _edit_locks:
+        _edit_locks[resolved] = asyncio.Lock()
+    lock = _edit_locks[resolved]
+    async with lock:
+        try:
+            with open(resolved, encoding="utf-8") as f:
+                content = f.read()
+        except FileNotFoundError:
+            return _mcp(f"File not found: {file_path}", error=True)
+        except PermissionError:
+            return _mcp(f"Permission denied: {file_path}", error=True)
+        except Exception as exc:
+            return _mcp(f"Failed to read {file_path}: {exc}", error=True)
+
+        count = content.count(old_string)
+        if count == 0:
+            return _mcp(f"old_string not found in {file_path}", error=True)
+        if count > 1 and not replace_all:
+            return _mcp(
+                f"old_string appears {count} times in {file_path}. "
+                "Use replace_all=true or provide a more unique string.",
+                error=True,
+            )
+
+        updated = (
+            content.replace(old_string, new_string)
+            if replace_all
+            else content.replace(old_string, new_string, 1)
         )
 
-    updated = (
-        content.replace(old_string, new_string)
-        if replace_all
-        else content.replace(old_string, new_string, 1)
-    )
-    try:
-        await _sandbox_write(sandbox, remote, updated)
-    except Exception as exc:
-        return _mcp(f"Failed to write {remote}: {exc}", error=True)
+        try:
+            with open(resolved, "w", encoding="utf-8") as f:
+                f.write(updated)
+        except Exception as exc:
+            return _mcp(f"Failed to write {file_path}: {exc}", error=True)
 
-    return _mcp(f"Edited {remote} ({count} replacement{'s' if count > 1 else ''})")
+    # Evict lock when no other coroutine is waiting, preventing unbounded growth.
+    if not lock.locked() and _edit_locks.get(resolved) is lock:
+        _edit_locks.pop(resolved, None)
+
+    return _mcp(f"Edited {resolved} ({count} replacement{'s' if count > 1 else ''})")
 
 
 async def _handle_glob(args: dict[str, Any]) -> dict[str, Any]:
@@ -608,3 +859,108 @@ E2B_FILE_TOOLS: list[tuple[str, str, dict[str, Any], Callable[..., Any]]] = [
 ]
 
 E2B_FILE_TOOL_NAMES: list[str] = [name for name, *_ in E2B_FILE_TOOLS]
+
+
+# ---------------------------------------------------------------------------
+# Unified tool descriptors — used by tool_adapter.py in both E2B and non-E2B modes
+# ---------------------------------------------------------------------------
+
+WRITE_TOOL_NAME = "Write"
+WRITE_TOOL_DESCRIPTION = (
+    "Write or create a file. Parent directories are created automatically. "
+    "For large content (>2000 words), prefer writing in sections using "
+    "bash_exec with 'cat > file' and 'cat >> file' instead."
+)
+WRITE_TOOL_SCHEMA: dict[str, Any] = {
+    "type": "object",
+    "properties": {
+        "file_path": {
+            "type": "string",
+            "description": (
+                "The path to the file to write. "
+                "Relative paths are resolved against the working directory."
+            ),
+        },
+        "content": {
+            "type": "string",
+            "description": "The content to write to the file.",
+        },
+    },
+}
+
+READ_TOOL_NAME = "read_file"
+READ_TOOL_DESCRIPTION = (
+    "Read a file from the working directory. Returns content with line numbers "
+    "(cat -n format). Use offset and limit to read specific ranges for large files."
+)
+READ_TOOL_SCHEMA: dict[str, Any] = {
+    "type": "object",
+    "properties": {
+        "file_path": {
+            "type": "string",
+            "description": (
+                "The path to the file to read. "
+                "Relative paths are resolved against the working directory."
+            ),
+        },
+        "offset": {
+            "type": "integer",
+            "description": (
+                "Line number to start reading from (0-indexed). Default: 0."
+            ),
+        },
+        "limit": {
+            "type": "integer",
+            "description": "Number of lines to read. Default: 2000.",
+        },
+    },
+}
+
+EDIT_TOOL_NAME = "Edit"
+EDIT_TOOL_DESCRIPTION = (
+    "Make targeted text replacements in a file. Finds old_string in the file "
+    "and replaces it with new_string. For replacing all occurrences, set "
+    "replace_all=true."
+)
+EDIT_TOOL_SCHEMA: dict[str, Any] = {
+    "type": "object",
+    "properties": {
+        "file_path": {
+            "type": "string",
+            "description": (
+                "The path to the file to edit. "
+                "Relative paths are resolved against the working directory."
+            ),
+        },
+        "old_string": {
+            "type": "string",
+            "description": "The text to find in the file.",
+        },
+        "new_string": {
+            "type": "string",
+            "description": "The replacement text.",
+        },
+        "replace_all": {
+            "type": "boolean",
+            "description": (
+                "Replace all occurrences of old_string (default: false). "
+                "When false, old_string must appear exactly once."
+            ),
+        },
+    },
+}
+
+
+def get_write_tool_handler(*, use_e2b: bool) -> Callable[..., Any]:
+    """Return the Write handler — both modes now use the unified _handle_write_file."""
+    return _handle_write_file
+
+
+def get_read_tool_handler(*, use_e2b: bool) -> Callable[..., Any]:
+    """Return the Read handler — both modes now use the unified _handle_read_file."""
+    return _handle_read_file
+
+
+def get_edit_tool_handler(*, use_e2b: bool) -> Callable[..., Any]:
+    """Return the Edit handler — both modes now use the unified _handle_edit_file."""
+    return _handle_edit_file
